+++ src/main/java/be/ugent/idlab/knows/functions/agent/functionModelProvider/fno/exception/PartialFunctionApplicationException.java
@@ -1,0 +1,7 @@
+package be.ugent.idlab.knows.functions.agent.functionModelProvider.fno.exception;
+
+public class PartialFunctionApplicationException extends FnOException{
+    public PartialFunctionApplicationException(String message) {
+        super(message);
+    }
+}